backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
mangum==0.17.0  # AWS Lambda adapter for FastAPI

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data validation
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# AI Services
openai==1.3.7
httpx==0.25.2  # For Ollama API calls
python-docx==1.1.0  # For DOCX resume parsing

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
hypothesis==6.92.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1
